store/wizard: move filesystem selectors to submodule
