+++ src/store/slices/wizard/filesystem/index.ts
@@ -1,3 +1,4 @@
+export * from './selectors';
 export { initialState as filesystemState } from './state';
 export * from './types';
 export * from './utilities';
